Move thread_name_lookup into dedicated module
